refactor(bt): continue phase4c step2 runtime module migration
